requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.14.0
keras==2.14.0
networkx==3.2
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
PyJWT==2.8.1
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
cryptography==41.0.7
requests==2.33.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
